+++ src/builder/ScssBuilder.ts
@@ -15,6 +15,7 @@ export interface ScssBuildConfig
 {
     entries: Record<string, string>;
     output: string;
+    hashFilenames: boolean;
     base: string;
     dependenciesMap: DependenciesMap;
 }
@@ -55,11 +56,13 @@ export class ScssBuilder
                 const filePath = path.join(buildConfig.base, buildConfig.entries[name]);
                 this.compilers[name] = new ScssCompiler(
                     this.logger,
+                    buildConfig.dependenciesMap,
                     buildConfig.base,
                     runConfig.debug,
                     buildConfig.output,
                     name,
                     filePath,
+                    buildConfig.hashFilenames
                 );
             }
         }